+++ src/app/components/Staking/Staking.tsx
@@ -21,7 +21,7 @@ import {
   GEO_BLOCK_MESSAGE,
   HealthCheckResult,
   HealthCheckStatus,
-} from "@/app/types/healthCheck";
+} from "@/app/types/services/healthCheck";
 import { getNetworkConfig } from "@/config/network.config";
 import {
   createStakingTx,
