+++ src/app/services/healthCheckService.ts
@@ -1,38 +1,24 @@
-import axios, { isAxiosError } from "axios";
-
+import { fetchHealthCheck, isAxiosError451 } from "../api/healthCheckClient";
 import {
   API_ERROR_MESSAGE,
   GEO_BLOCK_MESSAGE,
   HealthCheckResult,
   HealthCheckStatus,
-} from "../types/healthCheck";
-
-interface HealthCheckResponse {
-  data: string;
-}
+} from "../types/services/healthCheck";
 
 export const getHealthCheck = async (): Promise<HealthCheckResult> => {
   try {
-    const response = await axios.get(
-      `${process.env.NEXT_PUBLIC_API_URL}/healthcheck`,
-    );
-    const healthCheckAPIResponse: HealthCheckResponse = response.data;
-    // If the response has a data field, it's a normal response
+    const healthCheckAPIResponse = await fetchHealthCheck();
     if (healthCheckAPIResponse.data) {
       return {
         status: HealthCheckStatus.Normal,
         message: healthCheckAPIResponse.data,
       };
     } else {
-      // Something went wrong
       throw new Error(API_ERROR_MESSAGE);
     }
   } catch (error: Error | any) {
-    // Geo-blocking is a custom status code
-    if (
-      isAxiosError(error) &&
-      (error.response?.status === 451 || error.request.status === 451)
-    ) {
+    if (isAxiosError451(error)) {
       return {
         status: HealthCheckStatus.GeoBlocked,
         message: error.request?.response?.message || GEO_BLOCK_MESSAGE,
@@ -47,4 +33,8 @@ export const getHealthCheck = async (): Promise<HealthCheckResult> => {
       };
     }
   }
+};
+
+export const isGeoBlocked = (result: HealthCheckResult): boolean => {
+  return result.status === HealthCheckStatus.GeoBlocked;
 };
