+++ src/app/types/api/healthCheck.ts
@@ -1,0 +1,3 @@
+export interface HealthCheckResponse {
+  data: string;
+}
